backend-1/requirements.txt
asgiref==3.8.1
Django==5.1.3
django-cors-headers==4.6.0
djangorestframework==3.15.2
gunicorn==23.0.0
packaging==24.2
sqlparse==0.5.2
psycopg2-binary==2.9.9
python-decouple==3.8
typing_extensions==4.9.0
whitenoise==6.6.0
gevent==24.11.1
greenlet==3.1.1
pytz==2024.2
sqlparse==0.5.2
zope.event==5.0
zope.interface==7.2
